# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ClassIn": {
        "properties": {
          "capacity": {
            "exclusiveMinimum": 0.0,
            "maximum": 10000.0,
            "title": "Capacity",
            "type": "integer"
          },
          "level": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "capacity"
        ],
        "title": "ClassIn",
        "type": "object"
      },
...
      "EnrolledOut": {
        "properties": {
          "enrolled": {
            "title": "Enrolled",
            "type": "integer"
          }
        },
        "required": [
          "enrolled"
        ],
        "title": "EnrolledOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdNameOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "IdNameOut",
        "type": "object"
      },
...
      "SavedOut": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "SavedOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talent Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cohorts/{cohort_id}/classes": {
      "post": {
        "operationId": "create_class_api_v1_cohorts__cohort_id__classes_post",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdNameOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Class",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/composer/runs": {
      "get": {
        "operationId": "composer_runs_api_v1_cohorts__cohort_id__composer_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComposerRunOut"
                  },
                  "title": "Response Composer Runs Api V1 Cohorts  Cohort Id  Composer Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Composer Runs",
        "tags": [
          "cohorts"
        ]
      },
      "post": {
        "operationId": "composer_run_api_v1_cohorts__cohort_id__composer_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComposerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComposerRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Composer Run",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/enroll-accepted": {
      "post": {
        "operationId": "enroll_accepted_api_v1_cohorts__cohort_id__enroll_accepted_post",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrolledOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll Accepted",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/enrollments": {
      "get": {
        "operationId": "enrollments_api_v1_cohorts__cohort_id__enrollments_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(active|withdrawn|qualified|not_qualified)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "class_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Id"
            }
          },
          {
            "in": "query",
            "name": "track_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrollments",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/overview": {
      "get": {
        "operationId": "cohort_overview_api_v1_cohorts__cohort_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CohortOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cohort Overview",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/partner-demand": {
      "put": {
        "operationId": "set_demand_api_v1_cohorts__cohort_id__partner_demand_put",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Demand",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/qualification": {
      "get": {
        "operationId": "qualification_api_v1_cohorts__cohort_id__qualification_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QualificationRowOut"
                  },
                  "title": "Response Qualification Api V1 Cohorts  Cohort Id  Qualification Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qualification",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/stipends": {
      "get": {
        "operationId": "stipends_api_v1_cohorts__cohort_id__stipends_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StipendSummaryOut"
                  },
                  "title": "Response Stipends Api V1 Cohorts  Cohort Id  Stipends Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stipends",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/stipends/generate": {
      "post": {
        "operationId": "stipends_generate_api_v1_cohorts__cohort_id__stipends_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stipends Generate",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/stipends/pay": {
      "post": {
        "operationId": "stipends_pay_api_v1_cohorts__cohort_id__stipends_pay_post",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stipends Pay",
        "tags": [
          "cohorts"
        ]
      }
    },
    "/api/v1/cohorts/{cohort_id}/tracks/{track_id}": {
      "put": {
        "operationId": "set_track_capacity_api_v1_cohorts__cohort_id__tracks__track_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapacityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Track Capacity",
        "tags": [
          "cohorts"
        ]
      }
    },
...
  }
}